refactored footer styles
